requirements.txt
dnspython==2.6.1
requests==2.31.0
python-whois==0.9.4
